refactor(membership): added hooks for membership data
